feat(gitbbon-search): 모델 및 벡터 저장 정밀도를 fp32로 향상
- 모델 dtype: fp16 → fp32 (4개 파일)
- 벡터 저장: Int16 양자화 제거, Float32 직접 저장
- 검색 품질 향상을 위한 정밀도 개선

⚠️ 기존 벡터 캐시와 호환되지 않음, 재인덱싱 필요

+++ extensions/gitbbon-search/src/modelHost.ts
@@ -114,7 +114,7 @@ class ModelHost {
 			// "Expression produces a union type that is too complex to represent"
 			this.extractor = await (pipeline as Function)('feature-extraction', MODEL_NAME, {
 				device: useWebGPU ? 'webgpu' : 'wasm',
-				dtype: 'fp16',
+				dtype: 'fp32',
 				progress_callback: (p: { progress?: number; status?: string; file?: string }) => {
 					console.log('[ModelHost] Download progress:', p);
 					if (typeof p?.progress === 'number') {
