+++ extensions/gitbbon-search/src/webview/e5.worker.ts
@@ -51,7 +51,7 @@ async function initModel(): Promise<void> {
 		const useWebGPU = await checkWebGPU();
 		extractor = await pipeline('feature-extraction', MODEL_NAME, {
 			device: useWebGPU ? 'webgpu' : 'wasm',
-			dtype: 'fp16',
+			dtype: 'fp32',
 			// eslint-disable-next-line @typescript-eslint/no-explicit-any
 			progress_callback: (p: any) => {
 				if (typeof p?.progress === 'number') {
